third_party/WebKit/LayoutTests/http/tests/devtools/network/network-xhr-same-url-as-main-resource-expected.txt

CONSOLE MESSAGE: line 11: iframe loaded
CONSOLE MESSAGE: line 5: XHR loaded: 1
Tests that XHRs with the same url as a main resource have correct category.

Bug 68648  
xhr
